feat: support unbalanced block/statement in if
Instead of enforcing that both parts of an if-else-statement are
identical (either blocks or statements), update the grammar to identify
to which part the block or statement belongs.

This makes it possible to support one of them being a block and the
other being a statement. With that `else if` is supported, when using
blocks, as the `if` would be detected as a statement.

It also further simplifies the grammar for and while loops as they can
be for a block and statement too.